fix: keep output float32, only inputs float16

+++ sparse_matmul/train.py
@@ -162,12 +162,12 @@ def relu_sparse_Ax(vals, mask, shape, x):
     M, K = A_dense.shape
     N = x.shape[1]
 
-    out = torch.empty((M, N), device=x.device, dtype=torch.float16)
+    out = torch.empty((M, N), device=x.device, dtype=torch.float32)
 
     grid = lambda meta: (triton.cdiv(M, meta['BLOCK_M']), triton.cdiv(N, meta['BLOCK_N']))
     _matmul_relu_kernel[grid](A_dense, x.half(), out, M, N, K)
 
-    return out.float()
+    return out
 
 
 if __name__ == "__main__":
